refactor(PokemonLayout): add component to handle the pokemon layout in Home
