+++ frontend/src/components/PokemonLayout/PokemonLayout.module.css
@@ -1,0 +1,11 @@
+.pokedex {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  align-content: space-around;
+  gap: 50px 50px;
+}
+.title {
+  text-align: center;
+}
