+++ frontend/src/pages/Home/Home.module.css
@@ -1,12 +1,3 @@
-.pokedex {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  align-content: space-around;
-  gap: 50px 50px;
-}
-
 .title {
   text-align: center;
 }
